refactor test managers
