+++ tests/tt_metal/tt_metal/perf_microbenchmark/routing/tt_fabric_test_context.hpp
@@ -26,7 +26,8 @@
 #include "tt_fabric_test_memory_map.hpp"
 #include "tt_fabric_telemetry.hpp"
 #include "tt_fabric_test_progress_monitor.hpp"
-#include "tt_fabric_test_results.hpp"
+#include "tt_fabric_test_latency_results.hpp"
+#include "tt_fabric_test_bandwidth_results.hpp"
 #include "tt_fabric_test_bandwidth_profiler.hpp"
 #include "tt_fabric_test_eth_readback.hpp"
 #include "tt_fabric_test_code_profiler.hpp"
